Add modern layout selection
